Moved allPending logic functions to allPending in logic.
